Break out encode and decode machinery into SampleCoder.
Make all samples other than REPEAT samples be derived from ScalarSample.

Make all methods of SampleTimelineChunk be synchronized.

Implement the encoding on-the-fly using the lastSample mechanism, since
we only morph the lastSample.  This reduces the memory footprint of
accumulating samples by a factor of 8, and also provides, after
synchronizing the SampleTimelineChunk methods, thread-safe fetching of
compressed timelines as they are being built.