util: get rid of the TimelineRegistry
Solely use the CachingTimelineDAO instead. Under high load, we were
seeing a lot of MySQLIntegrityConstraintViolationException due to
synchronization issues between the DAO and the registry.

Since we already have a caching layer in place, move the dual lookup
functionality (hostId <-> hostName and simpleKindId <-> simpleKind)
into the CachingTimelineDAO and synchronize getOrAddHost and
getOrAddSampleKind. When the caches are warm, both method are equivalent
to a single lookup.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>