collector: try to downsize values on the way in
Don't necessarily trust the type of values coming in and try to save
space as early as possible on the way in.

Note that this is needed with the replayer as Jackson will try to do the
same by default and ScalarSample.compressSample performs unchecked
casts based on the opcode.

Note that this is not totally redundant with ScalarSample.compressSample
as the latter will perform further compression. This is mostly to handle
the replay scenario (where only basic type compression is performed).

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>