collector: major refactoring of timelines handling
We used to share a single timeline for all event types
(e.g. beans) for a given host.

This approach would work in practice if all hosts would report
all beans at the same time (e.g. using the BatchedEvent mechanism).
When regular Map events are used, each bean is reported serially.
Sharing a single timeline is then highly inefficient as the timeline
is filled with holes of nulls to align the samples received.

Instead, we now share a timeline only with the samples shared by a single
event type for a given host.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>