collector: fix aggregator tests
The aggregator dao layer was broken due to the recent merge.
Since I was going to update it anyways, I've ported it to use JDBI's
object API and StringTemplate 3.

Most of the founding work has been done to migrate the rest of the code
to the object API and StringTemplate.

One test is still failing:

testCache(com.ning.arecibo.collector.process.TestCollectorEventProcessor):
   In DecodedSampleOutputProcessor.getNextTime(), got -1 from timeCursor.getNextTime()

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>