collector: update /host_samples endpoint for the dashboard
Before, the endpoint was returning:

[ {
  "hostId" : 8,
  "sampleKindId" : 2900,
  "samplesAsCSV" : "..."
} ]

The id are useless for the dashboard. Instead, return:

[ {
  "hostName" : "box.company.com",
  "eventCategory" : "ActiveMQBroker_localhost",
  "sampleKind" : "TotalConsumerCount",
  "samples" : "..."
} ]

Note that we still pass CSV samples to the dashboard. We should change it to pass
the compressed timeline times and data ("compact view") to speed up data retrieval.
Most of the wiring is there and the decompression methods are shared (util submodule).

I've also updated the collector client to use the /host_samples endpoint exclusively.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>