dashboard: validate input dates
As reported by David, the dashboard should complain and not make a collector query
if the start time is greater than or equal to the end time.

This also adds unit tests for the Javascript code (via Jasmine).

Running tests on the command line uses Firefox 3 by default, which revealed
a bug in the dates manipulation.

As noted here:

  https://developer.mozilla.org/en/JavaScript/Reference/Global_Objects/Date/parse

Only starting with JavaScript 1.8.5 / Firefox 4, is a subset of ISO 8601 is supported.
We need to fall back to RFC2822 for the DateTimePicker widget, which is less user friendly,
but the dashboard will run on more browsers.

I've also commented a console.log() call, as console doesn't exist in Firefox unless Firebug
has been installed. We may not want to display this in production anyways.

Finally, to avoid firing Ajax requests when running tests, the UI setup functions are now
bound to custom events (content:loaded) instead of running as part of $(document).ready().

Reported-by: David Stryker <dstryker@ning.com>
Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>