dashboard: first pass at implementing super groups
Super groups are meant to build dashboards easily:
one can pre-configure in a configuration file (json format)
a list of beans and attributes to graph together.

In the dashboard, the group will be available as a node in
the sample kinds tree.

An example of configuration file is provided under
src/test/resources/groups.json.

Some UI work is required to make the groups stand out. It would be nice
to do it without creating a second sample kinds tree.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>