dashboard: switch loading indicator to a modal
Use a modal that covers the screen while data is loading.
This should make it more obvious to the user that the
dashboard is doing something and prevent him from further
clicking.

Note that nothing prevents him (while the modal is activated)
to press Esc. and/or click outside of the 'Data is loading...'
box to dismiss the modal. This won't stop the request in progress
though.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>