dashboard: UI refactoring, part 1
Refactored the JS code for hosts and sample kinds tree to use an
object-oriented (-ish) model. This makes it easier to keep state
thorough the actions.

Implemented etag mechanism for hosts and added lazy loading for the hosts tree.

This is only part 1 of the refactoring: tests need to be updated and
the /sample_kinds endpoint changed to add hosts information.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>